requirements.txt
streamlit
proxmoxer
requests
pandas
matplotlib
python-dotenv
watchdog
scikit-learn
joblib
